avoid rewriting .env if it already exists

+++ make.sh
@@ -14,17 +14,19 @@ case "${OSTYPE}" in
 		;;
 esac
 
-ORIGINAL_STTY="$(stty -g)"
-stty -echo
-cat > .env <<- EOE
-	USERID=${USERID}
-	GROUPID=${GROUPID}
-	USERLOGIN=${LOGNAME}
-	PASSWORD=$(mkpasswd --method=SHA-512 --stdin)
-	TIMEZONE=${TZ:-$(wget -O - -q http://geoip.ubuntu.com/lookup | sed 's#.*<TimeZone>\(.*\)</TimeZone>.*#\1#')}
-	I3THEME=${I3THEME:-default-light}
-EOE
-stty "${ORIGINAL_STTY}"
-echo
+if [ ! -e .env ]; then
+	ORIGINAL_STTY="$(stty -g)"
+	stty -echo
+	cat > .env <<- EOE
+		USERID=${USERID}
+		GROUPID=${GROUPID}
+		USERLOGIN=${LOGNAME}
+		PASSWORD=$(mkpasswd --method=SHA-512 --stdin)
+		TIMEZONE=${TZ:-$(wget -O - -q http://geoip.ubuntu.com/lookup | sed 's#.*<TimeZone>\(.*\)</TimeZone>.*#\1#')}
+		I3THEME=${I3THEME:-default-light}
+	EOE
+	stty "${ORIGINAL_STTY}"
+	echo
+fi
 
 DOCKER_BUILDKIT=1 COMPOSE_DOCKER_CLI_BUILD=1 docker-compose build
